[heroku-client] merged definitions in one declaration file
