--- types/heroku-client/url.d.ts
@@ -1,7 +1,0 @@
-export interface RequestURL {
-    host: string;
-    port: number;
-    secure: boolean;
-}
-
-export default function(url: string): RequestURL;
